+++ src/ext-lsv/module.make
@@ -1,2 +1,5 @@
 SRC += \
-    src/ext-lsv/lsvCmd.cpp
+    src/ext-lsv/lsvCmd.cpp \
+    src/ext-lsv/lsvPa1.cpp \
+    src/ext-lsv/lsvPa2.cpp \
+    src/ext-lsv/lsvUtils.cpp
